requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --pip-args=None requirements.in
#
amqp==5.3.1
    # via kombu
asgiref==3.9.2
    # via
    #   django
    #   django-cors-headers
asttokens==3.0.0
    # via stack-data
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
billiard==4.2.2
    # via celery
black==25.9.0
    # via -r requirements.in
celery==5.5.3
    # via -r requirements.in
certifi==2025.8.3
    # via sentry-sdk
click==8.3.0
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
coverage==7.10.7
    # via -r requirements.in
decorator==5.2.1
    # via ipython
dj-database-url==3.0.1
    # via -r requirements.in
django==5.2.6
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-cors-headers
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-cors-headers==4.9.0
    # via -r requirements.in
django-environ==0.12.0
    # via -r requirements.in
djangorestframework==3.16.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.5.1
    # via -r requirements.in
drf-spectacular==0.28.0
    # via -r requirements.in
executing==2.2.1
    # via stack-data
factory-boy==3.3.3
    # via -r requirements.in
faker==37.8.0
    # via factory-boy
inflection==0.5.1
    # via drf-spectacular
iniconfig==2.1.0
    # via pytest
ipython==9.5.0
    # via -r requirements.in
ipython-pygments-lexers==1.1.1
    # via ipython
isort==6.0.1
    # via -r requirements.in
jedi==0.19.2
    # via ipython
jsonschema==4.25.1
    # via drf-spectacular
jsonschema-specifications==2025.9.1
    # via jsonschema
kombu==5.5.4
    # via celery
matplotlib-inline==0.1.7
    # via ipython
mypy==1.18.2
    # via -r requirements.in
mypy-extensions==1.1.0
    # via
    #   black
    #   mypy
packaging==25.0
    # via
    #   black
    #   kombu
    #   pytest
parso==0.8.5
    # via jedi
pathspec==0.12.1
    # via
    #   black
    #   mypy
pexpect==4.9.0
    # via ipython
platformdirs==4.4.0
    # via black
pluggy==1.6.0
    # via pytest
prompt-toolkit==3.0.52
    # via
    #   click-repl
    #   ipython
psycopg[binary]==3.2.10
    # via -r requirements.in
psycopg-binary==3.2.10
    # via psycopg
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pygments==2.19.2
    # via
    #   ipython
    #   ipython-pygments-lexers
    #   pytest
pyjwt==2.10.1
    # via djangorestframework-simplejwt
pytest==8.4.2
    # via
    #   -r requirements.in
    #   pytest-django
pytest-django==4.11.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via celery
python-json-logger==3.3.0
    # via -r requirements.in
pytokens==0.1.10
    # via black
pyyaml==6.0.3
    # via drf-spectacular
redis==6.4.0
    # via -r requirements.in
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
ruff==0.13.2
    # via -r requirements.in
sentry-sdk==2.39.0
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
stack-data==0.6.3
    # via ipython
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.15.0
    # via mypy
tzdata==2025.2
    # via
    #   faker
    #   kombu
uritemplate==4.2.0
    # via drf-spectacular
urllib3==2.5.0
    # via sentry-sdk
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.14
    # via prompt-toolkit
whitenoise==6.11.0
    # via -r requirements.in
